// src/frontend/src/components/cron/types.ts
export interface Locale {
  everyText?: string
  emptyMonths?: string
  emptyMonthDays?: string
  emptyMonthDaysShort?: string
  emptyWeekDays?: string
  emptyWeekDaysShort?: string
  emptyHours?: string
  emptyMinutes?: string
  emptyMinutesForHourPeriod?: string
  yearOption?: string
  monthOption?: string
  weekOption?: string
  dayOption?: string
  hourOption?: string
  minuteOption?: string
  rebootOption?: string
  prefixPeriod?: string
  prefixMonths?: string
  prefixMonthDays?: string
  prefixWeekDays?: string
  prefixWeekDaysForMonthAndYearPeriod?: string
  prefixHours?: string
  prefixMinutes?: string
  prefixMinutesForHourPeriod?: string
  suffixMinutesForHourPeriod?: string
  errorInvalidCron?: string
  clearButtonText?: string
  weekDays?: string[]
  months?: string[]
  altWeekDays: string[]
  altMonths: string[]
}
// ...
export type PeriodType =
  | 'year'
  | 'month'
  | 'week'
  | 'day'
  | 'hour'
  | 'minute'
  | 'reboot'
export type AllowEmpty = 'always' | 'never' | 'for-default-value'
export type CronType =
  | 'period'
  | 'months'
  | 'month-days'
  | 'week-days'
  | 'hours'
  | 'minutes'
export type LeadingZeroType = 'month-days' | 'hours' | 'minutes'
export type LeadingZero = boolean | LeadingZeroType[]
export type ClockFormat = '24-hour-clock' | '12-hour-clock'
export type ShortcutsType =
  | '@yearly'
  | '@annually'
  | '@monthly'
  | '@weekly'
  | '@daily'
  | '@midnight'
  | '@hourly'
  | '@reboot'
export type Shortcuts = boolean | ShortcutsType[]
export type Mode = 'multiple' | 'single'
// ...
export interface FieldProps {
  value?: number[]
  setValue: SetValueNumbersOrUndefined
  locale: Locale
  className?: string
  disabled?: boolean
  readOnly?: boolean
  period: PeriodType
  periodicityOnDoubleClick?: boolean
  mode?: Mode
  allowClear?: boolean
  filterOption?: FilterOption
}
export interface PeriodProps
  extends Omit<
    FieldProps,
    | 'value'
    | 'setValue'
    | 'period'
    | 'periodicityOnDoubleClick'
    | 'mode'
    | 'filterOption'
  > {
  value: PeriodType
  setValue: SetValuePeriod
  shortcuts: Shortcuts
  allowedPeriods: PeriodType[]
}
// ...
export type SetValueNumbersOrUndefined = Dispatch<
  SetStateAction<number[] | undefined>
>
export type SetValuePeriod = Dispatch<SetStateAction<PeriodType>>
// ...
export type FilterOption = ({
  value,
  label,
}: {
  value: string
  label: string
}) => boolean